backend/requirements.txt
Django==5.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
drf-yasg==1.21.7
django-filter>=24.0
Pillow==10.2.0
python-decouple==3.8
setuptools<70  # Required for pkg_resources compatibility with Python 3.12
gunicorn==21.2.0  # Production WSGI server
whitenoise==6.6.0  # Static file serving

